create a task to freeze global hact
